Backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
pycryptodome==3.23.0
Pygments==2.19.2
PyPDF2==3.0.1
pytest==9.0.1
python-docx==1.2.0
python-dotenv==1.2.1
setuptools==80.9.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
wheel==0.45.1
